refactor : 파일 => 폴더로 자리 이동 <description> class인 파일들은 class폴더 안에, 다른 모듈들은 모두 modules폴더로 이동했다.
