--- src/init.js
@@ -1,6 +1,0 @@
-import {hideCounterElement, hideResultElement} from './elements/hideElement.js';
-
-export default function initGame() {
-  hideCounterElement();
-  hideResultElement();
-}
