+++ src/modules/init.js
@@ -1,0 +1,9 @@
+import {
+  hideCounterElement,
+  hideResultElement,
+} from '../elements/hideElement.js';
+
+export default function initGame() {
+  hideCounterElement();
+  hideResultElement();
+}
